namespace updates and orgnaization
