+++ Runtime/Prefabs/Framework/NGHexServices.prefab
@@ -1,0 +1,211 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1 &5983561296252206394
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 5358989190563841724}
+  - component: {fileID: 7981053519587688918}
+  m_Layer: 0
+  m_Name: NGHexServices
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &5358989190563841724
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5983561296252206394}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children:
+  - {fileID: 5910825261557439585}
+  - {fileID: 8509586077757001318}
+  m_Father: {fileID: 0}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &7981053519587688918
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5983561296252206394}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8475a3dac95f05c489e176ec40cd774d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ILog: {fileID: 5426957738340870119}
+  IMonoCallbacks: {fileID: 8509586077757001316}
+--- !u!1001 &2303541337791495898
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 5358989190563841724}
+    m_Modifications:
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6100055236613773629, guid: 61545384a3d29f34cacc32146bf8a808,
+        type: 3}
+      propertyPath: m_Name
+      value: NGHexLog
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 61545384a3d29f34cacc32146bf8a808, type: 3}
+--- !u!1 &5426957738340870119 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 6100055236613773629, guid: 61545384a3d29f34cacc32146bf8a808,
+    type: 3}
+  m_PrefabInstance: {fileID: 2303541337791495898}
+  m_PrefabAsset: {fileID: 0}
+--- !u!4 &5910825261557439585 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 5616176642045256379, guid: 61545384a3d29f34cacc32146bf8a808,
+    type: 3}
+  m_PrefabInstance: {fileID: 2303541337791495898}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1001 &5062178631374125228
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 5358989190563841724}
+    m_Modifications:
+    - target: {fileID: 3483648483795171016, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_Name
+      value: NGMonoCallbacks
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002, type: 3}
+--- !u!1 &8509586077757001316 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 3483648483795171016, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+    type: 3}
+  m_PrefabInstance: {fileID: 5062178631374125228}
+  m_PrefabAsset: {fileID: 0}
+--- !u!4 &8509586077757001318 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 3483648483795171018, guid: cd7fdbd47b3e77f4ebcd71f2ba5df002,
+    type: 3}
+  m_PrefabInstance: {fileID: 5062178631374125228}
+  m_PrefabAsset: {fileID: 0}
