+++ Runtime/Prefabs/Framework/NGMonoCallbacks.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &5983561296252206394
+--- !u!1 &3483648483795171016
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,22 +8,22 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 5358989190563841724}
-  - component: {fileID: 7981053519587688918}
+  - component: {fileID: 3483648483795171018}
+  - component: {fileID: 3483648483795171017}
   m_Layer: 0
-  m_Name: NGHexApp
+  m_Name: NGMonoCallbacks
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &5358989190563841724
+--- !u!4 &3483648483795171018
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5983561296252206394}
+  m_GameObject: {fileID: 3483648483795171016}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
@@ -31,16 +31,15 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &7981053519587688918
+--- !u!114 &3483648483795171017
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5983561296252206394}
+  m_GameObject: {fileID: 3483648483795171016}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8475a3dac95f05c489e176ec40cd774d, type: 3}
+  m_Script: {fileID: 11500000, guid: 5c4196c7ccf9429498ef80099e3221bf, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  ILog: {fileID: 6100055236613773629, guid: 61545384a3d29f34cacc32146bf8a808, type: 3}
