+++ Runtime/Prefabs/Framework/NGMonoCallbacks.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cd7fdbd47b3e77f4ebcd71f2ba5df002
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
