+++ Runtime/Scripts/Engine/Temporal/SimpleTimer.cs
@@ -2,6 +2,7 @@
 using UnityEngine;
 using HexCS.Core;
 using Event = HexCS.Core.Event;
+using HexUN.Framework.Services;
 
 namespace HexUN.Temporal
 {
