+++ Runtime/Scripts/Framework/Services/MiscServices/IMonoCallbacks.cs
@@ -1,0 +1,26 @@
+﻿using HexCS.Core;
+
+namespace HexUN.Framework.Services
+{
+    /// <summary>
+    /// Services providing global registration for basic 
+    /// monobehaviour lifecycle callbacks
+    /// </summary>
+    public interface IMonoCallbacks
+    {
+        /// <summary>
+        /// Subscriber for Update functions
+        /// </summary>
+        IEventSubscriber OnUpdate { get; }
+
+        /// <summary>
+        /// Subscriber for LateUpdate functions
+        /// </summary>
+        IEventSubscriber OnLateUpdate { get; }
+
+        /// <summary>
+        /// Subscriber for fixed update functions
+        /// </summary>
+        IEventSubscriber OnFixedUpdate { get; }
+    }
+}
