+++ Runtime/Scripts/Framework/Services/NGHexServices.cs.meta
