+++ Runtime/Scripts/Framework/SharedResource/ASOSharedResource.cs
@@ -1,5 +1,6 @@
 ﻿using HexCS.Core;
 
+using HexUN.Framework.Services;
 using HexUN.MonoB;
 
 using UnityEngine;
